+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ### Unreleased
 
 - feat!: React 19
+- feat!: Replace `eslint-plugin-react`with `@eslint-react/eslint-plugin`.
+- chore: Vendor the `ignored` prettier parser as `prettier-ignored-plugin.js` and drop the `prettier-plugin-ignored`
+  dependency.
 
 ### [2.0.0] - 2026-08-12